larger bandwidth for the TX2 to the board

+++ 5_task/inc/can_communication_task.h
@@ -21,9 +21,9 @@ typedef struct{
 } Gimbal_Send_Dbus_canStruct;
 
 typedef struct {
-    float vx;
-    float vy;
-    float vz;
+    double vx;
+    double vy;
+    double vz;
 } ROS_Msg_Struct;
 
 #ifdef __cplusplus
